+++ DlssUpdater/GameLibrary/EpicGames/CatalogCache.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.Json.Serialization;
+using System.Threading.Tasks;
+
+namespace DLSSUpdater.GameLibrary.EpicGames
+{
+    internal class CatalogCache
+    {
+        [JsonPropertyName("id")]
+        public string Id { get; set; } = string.Empty;
+
+        [JsonPropertyName("keyImages")]
+        public List<CachedGame> KeyImages { get; set; } = [];
+
+
+        internal class CachedGame
+        {
+            [JsonPropertyName("type")]
+            public string Type { get; set; } = string.Empty;
+
+            [JsonPropertyName("url")]
+            public string Url { get; set; } = string.Empty;
+        }
+    }
+}