+++ DlssUpdater/GameLibrary/ILibrary.cs
@@ -1,6 +1,8 @@
 ﻿using System.ComponentModel;
 using DlssUpdater.GameLibrary.Steam;
 using DLSSUpdater.Defines;
+using DLSSUpdater.GameLibrary;
+using DLSSUpdater.GameLibrary.EpicGames;
 using GameInfo = DlssUpdater.Defines.GameInfo;
 
 namespace DlssUpdater.GameLibrary;
@@ -11,6 +13,7 @@ public enum LibraryType
 
     Steam,
     Ubisoft,
+    EpicGames,
     GOG,
     // TODO: More launchers
 }
@@ -29,6 +32,7 @@ public interface ILibrary
             LibraryType.Steam => new SteamLibrary(config, logger),
             LibraryType.Ubisoft => new UbisoftConnectLibrary(config, logger),
             LibraryType.GOG => new GOGLibrary(config, logger),
+            LibraryType.EpicGames => new EpicGamesLibrary(config, logger),
             _ => throw new InvalidEnumArgumentException(nameof(config.LibraryType), (int)config.LibraryType, typeof(LibraryType))
         };
     }
@@ -41,6 +45,7 @@ public interface ILibrary
             LibraryType.Steam => "Steam",
             LibraryType.Ubisoft => "Ubisoft Connect",
             LibraryType.GOG => "GOG",
+            LibraryType.EpicGames => "Epic Games",
             _ => throw new InvalidEnumArgumentException(nameof(type), (int)type, typeof(LibraryType))
         };
     }